+++ CHANGELOG.md
@@ -10,6 +10,18 @@ You may also reference live-docs based on the openapi specification here:
 https://tapis-project.github.io/live-docs
 
 ---------------------------------------------------------------------------
+## 1.2.5 - 2022-??-??
+
+Incremental improvements and new preview features.
+
+### New features:
+- Support having a dynamic *rootDir* property. Attribute *rootDir* may start with *HOST_EVAL()*.
+- Update getSystem query parameter *resolveEffectUser*. Rename it *resolveEffective* and apply it to *effectiveUserId* and *rootDir*.
+
+### Bug fixes:
+- None.
+
+---------------------------------------------------------------------------
 ## 1.2.4 - 2022-08-07
 
 Incremental improvements and new preview features.